touccan_backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.0
packaging==25.0
Pillow==10.0.1
PyJWT==2.10.1
sqlparse==0.5.3